scripts/requirements-bridge.txt
# Serial → WebSocket bridge: pip install -r scripts/requirements-bridge.txt
pyserial>=3.5
websockets>=12.0
